backend/requirements.txt
fastapi
uvicorn
python-multipart
pypdf
langdetect
sentence-transformers
faiss-cpu
numpy